requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cffi==1.16.0
channels==4.1.0
crispy-bootstrap5==2024.2
cryptography==42.0.7
Django==4.2.13
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-browser-reload==1.12.1
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-crontab==0.7.1
django-excel-response==2.0.5
django-fontawesome-5==1.0.18
djangorestframework==3.14.0
et-xmlfile==1.1.0
fontawesomefree==6.5.1
jdcal==1.4.1
openpyxl==2.4.11
pycparser==2.22
PyMySQL==1.1.0
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
